Add confirm modal and service

+++ src/views/entry/overview.svelte
@@ -12,6 +12,7 @@
     sdk,
     StoreService,
     NotificationService,
+    ConfirmService,
   } from '../../services';
   import {
     Layout,
@@ -20,7 +21,6 @@
     OverflowMenu,
     OverflowMenuItem,
     EntryFullModelModal,
-    ConfirmDeleteModal,
     EntryFilterComponent,
   } from '../../components';
 
@@ -92,7 +92,6 @@
   let entryInFocus: EntryLiteModified;
   let languages: Language[] = [];
   let language: Language;
-  let entryToRemove: string = undefined;
 
   async function getEntries(
     _entries?: EntryLiteModified[],
@@ -169,17 +168,24 @@
   }
 
   async function removeEntry(id: string) {
-    await GeneralService.errorWrapper(
-      async () => {
-        await sdk.entry.deleteById({
-          id,
-          templateId,
-        });
-      },
-      async () => {
-        StoreService.update('entry', await sdk.entry.getAllLite(templateId));
-      }
-    );
+    if (
+      await ConfirmService.confirm(
+        'Delete Entry',
+        `Are you sure you want to delete this entry?`
+      )
+    ) {
+      await GeneralService.errorWrapper(
+        async () => {
+          await sdk.entry.deleteById({
+            id,
+            templateId,
+          });
+        },
+        async () => {
+          StoreService.update('entry', await sdk.entry.getAllLite(templateId));
+        }
+      );
+    }
   }
 
   function selectLanguage(id: string) {
@@ -313,8 +319,9 @@
                     text="Remove"
                     icon="trash"
                     on:click={() => {
-                      entryToRemove = entryLiteModified._id;
-                      StoreService.update('ConfirmDeleteModal', true);
+                      removeEntry(entryLiteModified._id);
+                      // entryToRemove = entryLiteModified._id;
+                      // StoreService.update('ConfirmDeleteModal', true);
                     }} />
                 </OverflowMenu>
               </li>
@@ -334,12 +341,4 @@
   <EntryFullModelModal
     entryId={entryInFocus ? entryInFocus._id : ''}
     templateId={template ? template._id : ''} />
-  <ConfirmDeleteModal
-    on:done={() => {
-      removeEntry(entryToRemove);
-      entryToRemove = undefined;
-    }}
-    on:cancel={() => {
-      entryToRemove = undefined;
-    }} />
 </Layout>
